backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
celery[redis]==5.4.0
boto3==1.36.4
httpx==0.28.1
pydantic-settings==2.7.1
anthropic==0.43.0
openai==1.59.7
mistralai==1.5.0
pgvector==0.3.6
python-dotenv==1.0.1
pytesseract==0.3.13
Pillow==11.3.0
pdf2image==1.17.0
pymupdf==1.26.5
minio==7.2.20
openpyxl==3.1.5
langgraph==0.6.11
langchain-core==0.3.83
